+++ src/frontend/ui/indicateur-métier/nombre-passage-urgence/ContenuNombreDePassagesAuxUrgences.tsx
